requirements.txt
fastmcp==2.11.3
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.28.1
python-multipart==0.0.9
pydantic==2.11.7
pypdf==6.0.0
python-docx==1.1.2
pydantic-settings>=2.11.7
